+++ src/main/kotlin/us/frollo/frollosdk/model/api/affordability/FPAssetLiabilityBreakdown.kt
@@ -24,8 +24,6 @@ data class FPAssetLiabilityBreakdown(
     @SerializedName("name") val name: String?,
     /**  The closing balance of liabilities breakdown; optional */
     @SerializedName("closing_balance") val closingBalance: BigDecimal?,
-    /**  The percentage of liabilities breakdown; optional */
-    @SerializedName("percentage") val percentage: Double?,
     /**  A list of accounts in assets or liabilities breakdown; optional */
     @SerializedName("accounts") val accounts: List<FPAssetsLiabilitiesBreakdownAccounts>?
 )