+++ src/main/kotlin/us/frollo/frollosdk/model/api/affordability/FPIncomeExpensesBreakdown.kt
@@ -19,18 +19,17 @@ import com.google.gson.annotations.SerializedName
 /** Data representation of the breakdown of Income or expenses of the Financial Passport */
 data class FPIncomeExpensesBreakdown(
 
-    /**  The category in this breakdown; Optional */
+    /** The category in this breakdown (Optional) */
     @SerializedName("category") val category: FPIncomeExpenseBreakdownCategory?,
-    /**  Number of transactions in this breakdown; Optional */
-    @SerializedName("number_of_transactions") val numberOfTransactions: Long?,
-    /**  The date of the first Transaction in this breakdown; Optional  . See [FinancialPassportResponse.DATE_FORMAT_PATTERN]*/
+    /** The date of the first Transaction in this breakdown. See [FinancialPassportResponse.DATE_FORMAT_PATTERN]; (Optional) */
     @SerializedName("first_transaction_date") val firstTransactionDate: String?,
-    /**  The date of the last Transaction in this breakdown; Optional  . See [FinancialPassportResponse.DATE_FORMAT_PATTERN]*/
+    /** The date of the last Transaction in this breakdown. See [FinancialPassportResponse.DATE_FORMAT_PATTERN]; (Optional) */
     @SerializedName("last_transaction_date") val lastTransactionDate: String?,
-    /**  An object to represent average of this breakdown; Optional */
+    /** An object to represent average of this breakdown (Optional) */
     @SerializedName("averages") val averages: FPIncomeExpenseAverages?,
-    /**  The [FPIncomeExpenseBreakdownTotals] in this breakdown; Optional */
-    @SerializedName("totals") val FPIncomeExpenseBreakdownTotals: FPIncomeExpenseBreakdownTotals?,
-    /**  An array of all Transactions that make up this expense breakdown. These can be used to query the Get Transactions API for more details on the underlying Transactions. */
+    /**
+     * An array of all Transactions that make up this expense breakdown.
+     * These can be used to query the Get Transactions API for more details on the underlying Transactions. (Optional)
+     */
     @SerializedName("transaction_ids") val transactionIds: List<Long>?
 )